requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.4
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.104.1
fastapi-utils==0.2.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.10.1
pymongo==4.6.0
pytest==8.0.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.24.0.post1
